+++ tests/testthat/test-estimate_ite_poisson.R
@@ -28,15 +28,14 @@ test_that("Poisson ITE Estimated Correctly", {
 
   # Poisson + Offset
   include_offset <- TRUE
-  offset_name <- "x1"
+  offset_name <- "x6"
   ite_result <- estimate_ite_poisson(y, z, X, X_names, include_offset,
                                      offset_name)
   expect_true(class(ite_result) == "numeric")
   expect_true(length(ite_result) == length(y))
 
   # Reproducible results
-  #expect_equal(ite_result[[1]], -0.8168112, tolerance = 0.000001)
-  #expect_equal(ite_result[[20]], -0.2781798, tolerance = 0.000001)
-  #expect_equal(ite_result[[91]], -0.2147065, tolerance = 0.000001)
-
+  expect_equal(ite_result[[1]], 0.002221956, tolerance = 0.000001)
+  expect_equal(ite_result[[19]], -0.05229289, tolerance = 0.000001)
+  expect_equal(ite_result[[72]], -0.5921032, tolerance = 0.000001)
 })