+++ tests/testthat/test-prune_rules.R
@@ -39,9 +39,9 @@ test_that("Rules Pruned Correctly", {
   ite <- ite_list[["ite"]]
   ite_std <- ite_list[["ite_std"]]
 
-  expect_equal(ite[10], -2.586544, tolerance = 0.000001)
-  expect_equal(ite[25], -1.553539, tolerance = 0.000001)
-  expect_equal(ite[70], 1.974034, tolerance = 0.000001)
+  expect_equal(ite[10], -4.876697, tolerance = 0.000001)
+  expect_equal(ite[25], -0.9431929, tolerance = 0.000001)
+  expect_equal(ite[70], 7.778598, tolerance = 0.000001)
 
 
   # Set parameters
@@ -73,9 +73,9 @@ test_that("Rules Pruned Correctly", {
   expect_equal(length(treelist),2)
   expect_equal(length(treelist[2]$list),100)
   expect_equal(colnames(treelist[2]$list[[1]])[1], "left daughter")
-  expect_equal(treelist[2]$list[[1]][2,6], 0.1122061, tolerance = 0.000001)
-  expect_equal(treelist[2]$list[[2]][3,6], -0.02796966, tolerance = 0.000001)
-  expect_equal(treelist[2]$list[[10]][3,6], -0.750918, tolerance = 0.000001)
+  expect_equal(treelist[2]$list[[1]][2,6], 0.6669895, tolerance = 0.000001)
+  expect_equal(treelist[2]$list[[2]][3,6], -0.3675594, tolerance = 0.000001)
+  expect_equal(treelist[2]$list[[10]][3,6], -1.344871, tolerance = 0.000001)
 
   rules <- extract_rules(treelist, X, ntrees, max_depth)
 
@@ -95,6 +95,6 @@ test_that("Rules Pruned Correctly", {
   # Correct outputs
   rules_RF <- prune_rules(rules, X, ite_std, max_decay, type_decay)
   expect_true(class(rules_RF) == "character")
-  #expect_equal(length(rules_RF), 190)
-  #expect_equal(rules_RF[3], "X[,3]>0.878150784662752")
+  #expect_equal(length(rules_RF), 155)
+  #expect_equal(rules_RF[3], "X[,3]<=0.5 & X[,8]>0.39 & X[,9]>0.1")
 })