+++ tests/testthat/test-split_data.R
@@ -36,8 +36,8 @@ test_that("split_data works as expected.", {
   z_inf <- inference[,2]
   X_inf <- inference[,3:ncol(inference)]
 
-  expect_equal(y_dis[2], 1.775837, tolerance = 0.00001)
+  expect_equal(y_dis[2], 1.842319, tolerance = 0.00001)
   expect_equal(z_inf[10], 1, tolerance = 0.00001)
-  expect_equal(X_dis[3,9][[1]], 1)
+  expect_equal(X_dis[3,9][[1]], 0.9244026, tolerance = 0.00001)
 
 })